Added automated setting of default grade boundaries when data is uploaded.
- When files are uploaded, the default/natural grade boundaries are
calculated, which serve to anchor a user and hopefully ensure they do
not move the sliders too far.
- It also lets them know the natural distribution of grades if they had
just followed UoA grading standards.
- This is done by taking the locations of students that have a raw
percentage of +- 5% around the grade boundary in UoA terms, and
averaging the locations.
- Not just taking students who are at 80% (for example) for the A/B
boundary, because a test might have a number of questions that’s not
divisible by 5 and so no-one gets 80%.